Handle False bool parameters correctly
This fixes a bug that I noticed with Streamlit 1.29.0 (including with
https://stonier-streamlit-parameters-demo-devel-rbcsdl.streamlit.app/?foo=False)
where if you uncheck a checkbox or otherwise set the query string
parameter for a bool to `False`, then the checkbox is checked.

You can see this quite easily by visiting
https://stonier-streamlit-parameters-demo-devel-rbcsdl.streamlit.app/?foo=False
Note that in the URL, `foo` is set to `False` but when this page loads,
the `Foo` checkbox will be checked. This fixes that bug by only setting
the parameter value to true when `s.lower() in ["true", "yes"]`.